ngpy-accounts-manager-api/requirements.txt
bcrypt==3.1.4
flask-restful==0.3.6
SQLAlchemy==1.3.7
object-mapper==1.0.7
Flask-Cors==3.0.3
gunicorn==19.7
psycopg2-binary==2.7.6.1
sendgrid==5.6.0
pytest==2.9.1
alembic==1.3.1
pyyaml==5.3.1
